+++ src/core/factories/analyzerFactory.ts
@@ -2,7 +2,6 @@ import { IAnalyzer } from "@interfaces/analyzer.interface";
 import { JSDeepAnalyzer } from "@languages/js/analyzers/jsDeepAnalyzer";
 import { PyDeepAnalyzer } from "@languages/python/analyzers/pyDeepAnalyzer";
 import { TSDeepAnalyzer } from "@languages/ts/analyzers/tsDeepAnalyzer";
-import chalk from "chalk";
 
 export class AnalyzerFactory {
   /**
@@ -22,8 +21,7 @@ export class AnalyzerFactory {
     }
 
     if (extension === "py") {
-      console.log(chalk.red("Python deep analyzer not implemented yet"));
-      // return new PyDeepAnalyzer();
+      return new PyDeepAnalyzer();
     }
 
     return null;